[diffusion] Move weight lock regression to diffusion tests
